# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
  /api/v1/compensation_changes:
    get:
      tags:
        - CompensationChanges
      summary: List CompensationChanges
      description: Get a paginated list of CompensationChanges with optional filtering and sorting
      operationId: listCompensationChanges
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompensationChangeList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - CompensationChanges
      summary: Create CompensationChange
      description: Create a new CompensationChange
      operationId: createCompensationChange
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCompensationChangeRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompensationChange'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/payroll_entries:
    get:
      tags:
        - PayrollEntries
      summary: List PayrollEntries
      description: Get a paginated list of PayrollEntries with optional filtering and sorting
      operationId: listPayrollEntries
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PayrollEntryList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - PayrollEntries
      summary: Create PayrollEntry
      description: Create a new PayrollEntry
      operationId: createPayrollEntry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePayrollEntryRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PayrollEntry'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
    CompensationChange:
      type: object
      required:
        - id
        - company_id
        - employee_id
        - change_type
        - metadata
      properties:
        id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        employee_id:
          type: string
          format: uuid
        change_type:
          type: string
        new_amount:
          nullable: true
          type: string
        effective_date:
          nullable: true
          type: string
          format: date
        reference_id:
          nullable: true
          type: string
          format: uuid
        note:
          nullable: true
          type: string
        metadata:
          type: object
...
    CreateCompensationChangeRequest:
      type: object
      required:
        - company_id
        - employee_id
        - change_type
        - metadata
      properties:
        company_id:
          type: string
          format: uuid
        employee_id:
          type: string
          format: uuid
        change_type:
          type: string
        new_amount:
          nullable: true
          type: string
        effective_date:
          nullable: true
          type: string
          format: date
        reference_id:
          nullable: true
          type: string
          format: uuid
        note:
          nullable: true
          type: string
        metadata:
          type: object
...
    CompensationChangeList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/CompensationChange'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    PayrollEntry:
      type: object
      required:
        - id
        - company_id
        - period_year
        - period_month
        - status
        - total_gross
        - total_deductions
        - total_net
        - metadata
      properties:
        id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        period_year:
          type: integer
          format: int32
        period_month:
          type: integer
          format: int32
        posting_date:
          nullable: true
          type: string
          format: date-time
        status:
          type: string
        salary_expense_account_id:
          nullable: true
          type: string
          format: uuid
        salary_payable_account_id:
          nullable: true
          type: string
          format: uuid
        total_gross:
          type: string
        total_deductions:
          type: string
        total_net:
          type: string
        journal_id:
          nullable: true
          type: string
          format: uuid
        accounting_post_id:
          nullable: true
          type: string
          format: uuid
        metadata:
          type: object
...
    CreatePayrollEntryRequest:
      type: object
      required:
        - company_id
        - period_year
        - period_month
        - status
        - total_gross
        - total_deductions
        - total_net
        - metadata
      properties:
        company_id:
          type: string
          format: uuid
        period_year:
          type: integer
          format: int32
        period_month:
          type: integer
          format: int32
        posting_date:
          nullable: true
          type: string
          format: date-time
        status:
          type: string
        salary_expense_account_id:
          nullable: true
          type: string
          format: uuid
        salary_payable_account_id:
          nullable: true
          type: string
          format: uuid
        total_gross:
          type: string
        total_deductions:
          type: string
        total_net:
          type: string
        journal_id:
          nullable: true
          type: string
          format: uuid
        accounting_post_id:
          nullable: true
          type: string
          format: uuid
        metadata:
          type: object
...
    PayrollEntryList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/PayrollEntry'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []